+++ SamplesApp/SamplesApp/Features/SwiftUIVideoChat/SwiftUIVideoChatEnvironment.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+enum SwiftUIVideoChatEnvironment {
+  static var urls: [URL] { Environment.urls }
+  static var channelId: String { Environment.channelId }
+  static var signalingConnectMetadata: Encodable? {
+    Environment.signalingConnectMetadata
+  }
+}